% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_factors.R
\name{armonizar_oficio}
\alias{armonizar_oficio}
\title{Armoniza los codigos de oficio}
\usage{
armonizar_oficio(d)
}
\arguments{
\item{d}{una encuesta CASEN en formato tibble o data.frame}
}
\value{
un tibble con la columna de oficio transformada
}
\description{
Convierte las etiquetas de los oficios para usar de manera 
uniforme los nombres de CASEN 2017, debiendo fusionar algunas categorias
que se intersectan (ejemplo: "Directivos de la Adm. Publica y Empresas" con
"Miembros del Poder Ejecuivo, Legislativo y Directivos de la Adm. Publica").
El procedimiento consiste en buscar la columna \code{oficio} (antes del anio 
2011) u \code{oficio1} (desde el anio 2011) y aplicar expresiones regulares 
de acuerdo a la codificacion respectiva.
}
\examples{
\dontrun{
casen1990 <- readRDS("casen1990.rds")
armonizar_oficio(casen1990)
}
}
