% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_region.R
\name{armonizar_regiones}
\alias{armonizar_regiones}
\title{Armoniza los codigos de regiones}
\usage{
armonizar_regiones(d)
}
\arguments{
\item{d}{una encuesta CASEN en formato tibble o data.frame}
}
\value{
un tibble con la columna de region transformada
}
\description{
Convierte las etiquetas de las regiones para usar de manera 
uniforme los nombres de CASEN 2017. El procedimiento consiste en buscar la 
columna \code{r} (antes del anio 2000) o \code{region} (desde el anio 2000) 
y aplicar expresiones regulares de acuerdo a la codificacion respectiva.
}
