\name{casen_2017_los_rios}
\alias{casen_2017_los_rios}
\docType{data}
\title{
Conjunto CASEN 2017
}
\description{
Datos de demostracion limitados a las comunas de la Region de Los Rios con
representatividad comunal (Valdivia, Los Rios, Paillaco, Panguipulli,
La Union y Río Bueno).
}
\usage{data("casen_2017_los_rios")}
\format{
  Un data frame con 8055 observaciones en las siguientes 7 variables.
  \describe{
    \item{\code{expc}}{factor de expansion comunal}
    \item{\code{varstrat}}{estratos de varianza}
    \item{\code{varunit}}{conglomerados de varianza}
    \item{\code{ytotcorh}}{ingreso total del hogar corregido}
    \item{\code{sexo}}{sexo}
    \item{\code{comuna}}{comuna}
    \item{\code{pobreza}}{situacion de pobreza por ingresos}
  }
}
\source{
Ministerio de Desarrollo Social (MDS)
}
\examples{
casen_2017_los_rios
}
\keyword{datasets}
