% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_input}
\alias{check_input}
\title{Check functions input}
\usage{
check_input(datos, variable, agrupacion, peso, conglomerado, estrato)
}
\arguments{
\item{datos}{un data.frame o tibble con la encuesta CASEN (o un subconjunto acotado a una region, etc)}

\item{variable}{una columna de tipo numerico, por ejemplo ytotcorh que es la opcion por defecto}

\item{agrupacion}{una columna de tipo texto/factor, por ejemplo region que es la opcion por defecto}

\item{peso}{una columna de tipo numerico, por defecto es expr que corresponde al factor de expansion regional de acuerdo al manual CASEN 2017}

\item{conglomerado}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}

\item{estrato}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}
}
\description{
Check functions input
}
\keyword{internal}
