% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_nas}
\alias{check_nas}
\title{Verifica la presencia de NAs en los intervalos de confianza}
\usage{
check_nas(d)
}
\arguments{
\item{d}{un data.frame o tibble con un estadistico agrupado}
}
\description{
Verifica la presencia de NAs en los intervalos de confianza
}
\keyword{internal}
