\name{codigos_subdere}
\alias{codigos_subdere}
\docType{data}
\title{
Codigos Unicos Territoriales (CUT)
}
\description{
Corresponde a los codigos oficiales de comunas.
}
\usage{codigos_subdere}
\format{
  Un tibble con 346 filas y las siguientes seis columnas.
  \describe{
    \item{\code{codigo_comuna}}{codigo oficial de comuna}
    \item{\code{nombre_comuna}}{nombre oficial de comuna}
    \item{\code{codigo_provincia}}{codigo oficial de provincia}
    \item{\code{nombre_provincia}}{nombre oficial de provincia}
    \item{\code{codigo_region}}{codigo oficial de region}
    \item{\code{nombre_region}}{nombre oficial de region}
  }
}
\source{
Elaboracion propia con base en datos de Subsecretaria de Desarrollo Regional (SUBDERE).
}
\examples{
codigos_subdere
}
\keyword{datasets}
