% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_casen.R
\name{descargar_casen}
\alias{descargar_casen}
\title{Download CASEN from MDS website}
\usage{
descargar_casen(anios = NULL, carpeta = getwd())
}
\arguments{
\item{anios}{si no se indica un anio, descarga todos los anios disponibles}

\item{carpeta}{si no se indica una carpeta, descarga en la carpeta de
trabajo o la raiz del projecto}
}
\value{
los archivos comprimidos de la encuesta CASEN descargados desde el
sitio web del MDS
}
\description{
Download CASEN from MDS website
}
\examples{
# descargar todas las encuestas disponibles
# descargar_casen()

# descargar CASEN 2017 en carpeta especifica
# descargar_casen(2017, "data-raw")
}
