% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_casen.R
\name{descargar_casen_github}
\alias{descargar_casen_github}
\title{Descarga la encuesta CASEN de GitHub}
\usage{
descargar_casen_github(anios = NULL, carpeta = getwd())
}
\arguments{
\item{anios}{si no se indica un anio, descarga todos los anios disponibles}

\item{carpeta}{si no se indica una carpeta, descarga en la carpeta de
trabajo}
}
\value{
los archivos rds de la encuesta CASEN descargados desde GitHub
}
\description{
Obtiene los archivos de la encuesta en formato R.
}
\examples{
# descargar todas las encuestas disponibles
# descargar_casen_github()

# descargar CASEN 2017 en carpeta especifica
# descargar_casen_github(2017, "data-raw")
}
