% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_casen.R
\name{descargar_casen_mds}
\alias{descargar_casen_mds}
\title{Descarga la encuesta CASEN del sitio web del Ministerio de Desarrollo Social}
\usage{
descargar_casen_mds(anios = NULL, carpeta = NULL)
}
\arguments{
\item{anios}{si no se indica un anio, descarga todos los anios disponibles}

\item{carpeta}{se debe especificar una carpeta de descarga}
}
\value{
los archivos comprimidos de la encuesta CASEN descargados desde el
sitio web del MDS
}
\description{
Obtiene los archivos de la encuesta en formato RAR o ZIP segun
el anio de la encuesta. No descomprime ni modifica los datasets originales.
}
\examples{
# descargar CASEN 1990 en carpeta temporal
descargar_casen_mds(1990, tempdir())
}
