% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{estadistica_descriptiva}
\alias{estadistica_descriptiva}
\title{Estadistica descriptiva usando disenio complejo}
\usage{
estadistica_descriptiva(datos, variable = "ytotcorh",
  agrupacion = "region", peso = "expr", cuantiles = 1:100/100,
  conglomerado = "varunit", estrato = "varstrat", solo_media = FALSE)
}
\arguments{
\item{datos}{un data.frame o tibble con la encuesta CASEN (o un subconjunto
acotado a una region, etc)}

\item{variable}{una columna de tipo numerico, por ejemplo ytotcorh que es la
opcion por defecto}

\item{agrupacion}{una columna de tipo texto/factor, por ejemplo region que es
la opcion por defecto}

\item{peso}{una columna de tipo numerico, por defecto es expr que corresponde
al factor de expansion regional de acuerdo al manual CASEN 2017}

\item{cuantiles}{un vector de tipo numerico, por defecto es 1:100 / 100}

\item{conglomerado}{una columna de tipo numerico, por defecto es varunit de
acuerdo al manual CASEN 2017}

\item{estrato}{una columna de tipo numerico, por defecto es varunit de
acuerdo al manual CASEN 2017}

\item{solo_media}{valor logico para omitir el calculo de mediana y cuantiles,
por defecto es FALSE}
}
\value{
una lista con tres data frames que contienen las medias, medianas y
percentiles
}
\description{
\code{estadistica_descriptiva} define el disenio a partir de los
conglomerados, estratos y pesos, con base en esto obtiene las medidas de
tendencia central
}
\examples{
r14 <- leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))
estadistica_descriptiva(r14, "ytotcorh", c("comuna", "sexo"), "expc", 0.7)

}
