% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{extract}
\alias{extract}
\title{Extract CASEN from zip/rar file}
\usage{
extract(archive, file = 1L, mode = "r", format = NULL,
  filter = NULL)
}
\arguments{
\item{mode}{how to open the connection as used in \code{base::connections()}}
}
\description{
Extract CASEN from zip/rar file
}
\keyword{internal}
