% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_casen.R
\name{leer_casen}
\alias{leer_casen}
\title{Lee los archivos descargados de la encuesta CASEN}
\usage{
leer_casen(archivo)
}
\arguments{
\item{archivo}{cadena de texto con la ruta de la encuesta en formato zip/rar o sav/dta}
}
\value{
la encuesta CASEN en formato tibble
}
\description{
Lee los archivos descargados de la encuesta CASEN
}
\examples{
# datos de ejemplo en el paquete
leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))

# descargando los datos
# descargar_casen(2017, "data-raw")
# leer_casen("data-raw/2017_spss.rar")
}
