% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{mean_median}
\alias{mean_median}
\title{Envolvente para estadistica descriptiva con diseños complejos}
\usage{
mean_median(
  d_groups,
  des,
  stat_fun,
  agrupacion,
  conglomerado,
  estrato,
  peso,
  variable,
  col_prefix
)
}
\arguments{
\item{d_groups}{un vector que sea la salida de unique_tuples()}

\item{des}{disenio de encuesta a partir de un data.frame o tibble}

\item{stat_fun}{la funcion a usar para la agregacion, por ejemplo srvyr::survey_mean()}

\item{agrupacion}{una columna de tipo texto/factor, por ejemplo region que es la opcion por defecto}

\item{conglomerado}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}

\item{estrato}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}

\item{peso}{una columna de tipo numerico, por defecto es expr que corresponde al factor de expansion regional de acuerdo al manual CASEN 2017}

\item{variable}{una columna de tipo numerico, por ejemplo ytotcorh que es la opcion por defecto}

\item{col_prefix}{prefijo para agregar a las columnas resultantes}
}
\value{
Una tabla con las medias o medianas agrupadas y su intervalo de confianza.
}
\description{
Envolvente para estadistica descriptiva con diseños complejos
}
\keyword{internal}
