% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{mediana_agrupada}
\alias{mediana_agrupada}
\title{Mediana agrupada usando diseño complejo}
\usage{
mediana_agrupada(disenio)
}
\arguments{
\item{disenio}{la salida de `configuracion_disenio()` que provee ademas los
grupos y las variables en forma de lista}
}
\value{
Una tabla con las medianas agrupadas y su intervalo de confianza.
}
\description{
Usa los factores de expansion, conglomerados y estratos para
calcular correctamente las medianas comunales o regionales.
}
\examples{
cd <- configuracion_disenio(casen_2017_los_rios, "ytotcorh", c("comuna", "sexo"), "expc")
mediana_agrupada(cd)
}
