% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{modelo_lineal_generalizado}
\alias{modelo_lineal_generalizado}
\title{Modelo lineal generalizado con disenio complejo}
\usage{
modelo_lineal_generalizado(datos, variable_dependiente = "ytotcorh",
  variables_independientes = "region", peso = "expr",
  conglomerado = "varunit", estrato = "varstrat")
}
\arguments{
\item{datos}{un data.frame o tibble con la encuesta CASEN (o un subconjunto acotado a una region, etc)}

\item{variable_dependiente}{una columna de tipo numerico, por ejemplo ytotcorh que es la opcion por defecto}

\item{variables_independientes}{una o mas columnas de tipo numerico o factor, por ejemplo region que es la opcion por defecto}

\item{peso}{una columna de tipo numerico, por defecto es expr que corresponde al factor de expansion regional de acuerdo al manual CASEN 2017}

\item{conglomerado}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}

\item{estrato}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}
}
\value{
un tibble con los resultados del modelo lineal generalizado (funcion glm + disenio complejo)
}
\description{
Modelo lineal generalizado con disenio complejo
}
\examples{
r14 <- leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))
modelo_lineal_generalizado(r14, "ytotcorh", c("comuna", "sexo"), "expc")
}
