% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{percentiles_agrupados}
\alias{percentiles_agrupados}
\title{Percentiles agrupados usando disenio complejo}
\usage{
percentiles_agrupados(datos, variable = "ytotcorh",
  agrupacion = "region", peso = "expr", percentiles = 0.7,
  conglomerado = "varunit", estrato = "varstrat")
}
\arguments{
\item{datos}{un data.frame o tibble con la encuesta CASEN (o un subconjunto acotado a una region, etc)}

\item{variable}{una columna de tipo numerico, por ejemplo ytotcorh que es la opcion por defecto}

\item{agrupacion}{una columna de tipo texto/factor, por ejemplo region que es la opcion por defecto}

\item{peso}{una columna de tipo numerico, por defecto es expr que corresponde al factor de expansion regional de acuerdo al manual CASEN 2017}

\item{percentiles}{percentiles a calcular, si no se especifica calcula el percentil 70}

\item{conglomerado}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}

\item{estrato}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}
}
\value{
una tabla con los percentiles y su error estandar
}
\description{
Percentiles agrupados usando disenio complejo
}
\examples{
r14 <- leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))
percentiles_agrupados(r14, "ytotcorh", c(0.25, 0.5, 0.75), c("comuna", "sexo"), "expc")
}
