% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_nas}
\alias{check_nas}
\title{Checks NAs in confidence intervals}
\usage{
check_nas(d)
}
\arguments{
\item{d}{un data.frame o tibble con un estadistico agrupado}
}
\description{
Checks NAs in confidence intervals
}
\keyword{internal}
