% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_casen.R
\name{descargar_casen_github}
\alias{descargar_casen_github}
\title{Descarga la encuesta CASEN de GitHub}
\usage{
descargar_casen_github(anios = NULL, carpeta = NULL)
}
\arguments{
\item{anios}{si no se indica un anio, descarga todos los anios disponibles}

\item{carpeta}{se debe especificar una carpeta de descarga}
}
\value{
los archivos rds de la encuesta CASEN descargados desde GitHub
}
\description{
Obtiene los archivos de la encuesta en formato R.
}
\examples{
# descargar CASEN 1990 en carpeta temporal
descargar_casen_github(1990, tempdir())
}
