% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_casen.R
\name{descargar_casen_mds}
\alias{descargar_casen_mds}
\title{Descarga la encuesta CASEN del sitio web del MDS}
\usage{
descargar_casen_mds(anios = NULL, carpeta = getwd())
}
\arguments{
\item{anios}{si no se indica un anio, descarga todos los anios disponibles}

\item{carpeta}{si no se indica una carpeta, descarga en la carpeta de
trabajo}
}
\value{
los archivos comprimidos de la encuesta CASEN descargados desde el
sitio web del MDS
}
\description{
Obtiene los archivos de la encuesta en formato RAR o ZIP segun
el anio de la encuesta. No descomprime ni modifica los datasets originales.
}
\examples{
# descargar todas las encuestas disponibles
# descargar_casen_mds()

# descargar CASEN 2017 en carpeta especifica
# descargar_casen_mds(2017, "data-raw")
}
