% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{mediana_agrupada}
\alias{mediana_agrupada}
\title{Mediana agrupada usando disenio complejo}
\usage{
mediana_agrupada(disenio)
}
\arguments{
\item{disenio}{la salida de `configuracion_disenio()` que provee ademas los
grupos y las variables en forma de lista}
}
\value{
una tabla con las medianas agrupadas y su intervalo de confianza
}
\description{
Mediana agrupada usando disenio complejo
}
\examples{
r14 <- leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))
cd <- configuracion_disenio(r14, "ytotcorh", c("comuna", "sexo"), "expc")
mediana_agrupada(cd)
}
