% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{mediana_agrupada}
\alias{mediana_agrupada}
\title{Mediana agrupada usando disenio complejo}
\usage{
mediana_agrupada(datos, variable = "ytotcorh", agrupacion = "region",
  peso = "expr", conglomerado = "varunit", estrato = "varstrat")
}
\arguments{
\item{datos}{un data.frame o tibble con la encuesta CASEN (o un subconjunto acotado a una region, etc)}

\item{variable}{una columna de tipo numerico, por ejemplo ytotcorh que es la opcion por defecto}

\item{agrupacion}{una columna de tipo texto/factor, por ejemplo region que es la opcion por defecto}

\item{peso}{una columna de tipo numerico, por defecto es expr que corresponde al factor de expansion regional de acuerdo al manual CASEN 2017}

\item{conglomerado}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}

\item{estrato}{una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017}
}
\value{
una tabla con las medianas agrupadas y su intervalo de confianza
}
\description{
Mediana agrupada usando disenio complejo
}
\examples{
r14 <- leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))
mediana_agrupada(r14, "ytotcorh", c("comuna", "sexo"), "expc")
}
