% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{modelo_lineal_generalizado}
\alias{modelo_lineal_generalizado}
\title{Modelo lineal generalizado con diseño complejo}
\usage{
modelo_lineal_generalizado(disenio, modelo = "ytotcorh ~ sexo")
}
\arguments{
\item{disenio}{la salida de `configuracion_disenio()` que provee además los
grupos y las variables en forma de lista}

\item{modelo}{una expresion simbólica escrita como cadena de texto, consulta
[stats::formula()] para los detalles de modelos con componentes logarítmicas,
cuadráticas, etc.}
}
\value{
una lista cuyas clases son svyglm, glm y lm
}
\description{
Modelo lineal generalizado con diseño complejo
}
\examples{
cd <- configuracion_disenio(casen_2017_los_rios, "ytotcorh", c("comuna", "sexo"), "expc")
modelo_lineal_generalizado(cd, "ytotcorh ~ comuna + sexo")

}
