% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{percentiles_agrupados}
\alias{percentiles_agrupados}
\title{Percentiles agrupados usando diseño complejo}
\usage{
percentiles_agrupados(disenio, percentiles = 0.7)
}
\arguments{
\item{disenio}{la salida de `configuracion_disenio()` que provee además los
grupos y las variables en forma de lista}

\item{percentiles}{percentiles a calcular, si no se especifica calcula el
percentil 70}
}
\value{
una tabla con los percentiles y su error estándar
}
\description{
Usa los factores de expansion, conglomerados y estratos para
calcular correctamente los percentiles comunales o regionales.
}
\examples{
cd <- configuracion_disenio(casen_2017_los_rios, "ytotcorh", c("comuna", "sexo"), "expc")
percentiles_agrupados(cd, 0.7)
}
