% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{unique_groups}
\alias{unique_groups}
\title{Obtiene los grupos únicos a partir de las variables de agrupación
Extract unique groups from grouping variables combinations}
\usage{
unique_groups(d, agrupacion)
}
\arguments{
\item{d}{un data.frame o tibble}

\item{agrupacion}{columnas de tipo teto/factor en d}
}
\description{
Obtiene los grupos únicos a partir de las variables de agrupación
Extract unique groups from grouping variables combinations
}
\keyword{internal}
